SentinelSweep-SOC/requirements.txt
python-dotenv>=1.0.0
pandas>=2.0.0
rich>=13.0.0
ipaddress>=1.0.23
requests>=2.31.0
